add ovs support for fakebridge.
This was request by a proxmox user on ifupdown2 github issues.

https://github.com/CumulusNetworks/ifupdown2/issues/245

It was possible to do it with ifupdown1, but not ifupdown2.

The patch is already applied upstream
https://github.com/CumulusNetworks/ifupdown2/pull/246

example to define a fake ovs bridge with vlan10:

auto vmbr0
iface vmbr0 inet manual
        ovs_type OVSBridge

auto vmbr0v10
iface vmbr0v10 inet manual
        ovs_type OVSBridge
        ovs_bridge vmbr0
        ovs_options vmbr0 10

Signed-off-by: Alexandre Derumier <aderumier@odiso.com>