patch: fix bond mac address at boot.
since systemd v241, like for bridge, the bond mac is setup randomly at
boot, instead inherit from first slave.

Then, on next ifreload, ifupdown2 was already fixing it, but with an
down/up of the bond (with potentials impact on the network).

This patch have been sent upstream:
https://github.com/CumulusNetworks/ifupdown2/pull/270

This patch is applying fine too for ifupdown2 for proxmox7

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>