fix #5009: avoid waiting for stdout eof of /etc/network/ scripts
Previously, PVE (7 and 8) hosts would hang at boot if both ntpsec and
ntpsec-ntpdate are installed. The root cause for the hang is an
unfortunate interaction between ntpsec, ntpsec-ntpdate and the PVE
ifupdown2 package. The hang happens because ntpsec-ntpdate installs a
hook /etc/network/if-up.d/ntpsec-ntpdate that blocks until networking
is available if ntpsec is installed. Previously, ifupdown2 would wait
for the hook to terminate, so networking would never become available,
resulting in a deadlock. See the bug report [0] for more information.

With this patch, ifupdown2 runs the hook in the background and does
not wait for it to terminate, thus resolving the deadlock.

This patch was already applied upstream [1]. Backport it to make it
available before the next upstream release.

[0] https://bugzilla.proxmox.com/show_bug.cgi?id=5009
[1] https://github.com/CumulusNetworks/ifupdown2/pull/274

Signed-off-by: Friedrich Weber <f.weber@proxmox.com>