d/maintscripts: use postinst arguments to determine first install
Following https://www.debian.org/doc/debian-policy/ap-flowcharts.html
postinst gets called as 'postinst configure' w/o second argument on
first installation, use that information instead of creating a
flag-file in preinst.

Technically this changes the semantics and the first-install parts
will not be run, in case ifupdown2 had been installed on the system,
but was subsequently removed, but not purged ('rc' in dpkg-output).
Since the functionality was added quite a while ago (released with PVE
6.2 to support ovs-setups with ifupdown2 - the potential for
regression should be limited. Originally introduced in:
849ae55de6454ea4631de0899182c148c954e46a

Fixes: 8d5303c35044b612afc68eec1dc5bc265a5dd328
Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>